+++ src/main/java/com/myweb/MusicLD/repository/PaymentRepository.java
@@ -4,7 +4,9 @@ import com.myweb.MusicLD.entity.PaymentEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 @Repository
-public interface PaymentRepository extends JpaRepository<PaymentEntity, Long> {
+public interface PaymentRepository extends JpaRepository<PaymentEntity, BigInteger> {
 
 }
